# golangapi/internal/modules/users/delivery/http/handlers.go
package http

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"
	"strings"

	"icmongolang/config"
	"icmongolang/internal/middleware"
	"icmongolang/internal/models"
	"icmongolang/internal/modules/users"
	"icmongolang/internal/modules/users/presenter"
	"icmongolang/pkg/httpErrors"
	"icmongolang/pkg/logger"
	"icmongolang/pkg/responses"
	"icmongolang/pkg/utils"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/render"
	"github.com/google/uuid"
)

type userHandler struct {
	cfg     *config.Config
	usersUC users.UserUseCaseI
	logger  logger.Logger
}

func CreateUserHandler(uc users.UserUseCaseI, cfg *config.Config, logger logger.Logger) users.Handlers {
	return &userHandler{cfg: cfg, usersUC: uc, logger: logger}
}

// SignInEmail – POST /api/auth/signin (email + password)
// @Summary      User login with email and password
// @Description  Authenticates a user using email and password. Returns access and refresh tokens.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        request body presenter.UserSignIn true "User credentials"
// @Success      200  {object}  map[string]interface{}  "Tokens returned successfully"
// @Failure      400  {object}  responses.ErrorResponse  "Bad request / validation error"
// @Failure      401  {object}  responses.ErrorResponse  "Unauthorized (wrong email/password)"
// @Failure      404  {object}  responses.ErrorResponse  "User not found"
// @Router       /auth/signin [post]
func (h *userHandler) SignInEmail() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		req := new(presenter.UserSignIn)
		contentType := r.Header.Get("Content-Type")

		var err error
		if strings.Contains(contentType, "application/json") {
			err = json.NewDecoder(r.Body).Decode(req)
		} else if strings.Contains(contentType, "multipart/form-data") {
			err = utils.DecodeFormToStruct(r, req)
		} else {
			err = errors.New("unsupported content type")
		}

		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}

		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}

		accessToken, refreshToken, err := h.usersUC.SignIn(r.Context(), req.Email, req.Password)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}

		render.Respond(w, r, responses.CreateSuccessResponse(presenter.Token{
			AccessToken:  accessToken,
			RefreshToken: refreshToken,
			TokenType:    "Bearer",
		}))
	}
}

// SignInUsername – POST /api/auth/login (username + password)
// @Summary      User login with username and password
// @Description  Authenticates a user using username and password. Returns access and refresh tokens.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        request body presenter.SignIn true "User credentials (username + password)"
// @Success      200  {object}  map[string]interface{}  "Tokens returned successfully"
// @Failure      400  {object}  responses.ErrorResponse  "Bad request / validation error"
// @Failure      401  {object}  responses.ErrorResponse  "Unauthorized (wrong username/password)"
// @Failure      404  {object}  responses.ErrorResponse  "User not found"
// @Router       /auth/login [post]
func (h *userHandler) SignInUsername() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		req := new(presenter.SignIn)
		contentType := r.Header.Get("Content-Type")

		var err error
		if strings.Contains(contentType, "application/json") {
			err = json.NewDecoder(r.Body).Decode(req)
		} else if strings.Contains(contentType, "multipart/form-data") {
			err = utils.DecodeFormToStruct(r, req)
		} else {
			err = errors.New("unsupported content type")
		}

		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}

		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}

		accessToken, refreshToken, err := h.usersUC.SignInByUsername(r.Context(), req.Username, req.Password)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}

		render.Respond(w, r, responses.CreateSuccessResponse(presenter.Token{
			AccessToken:  accessToken,
			RefreshToken: refreshToken,
			TokenType:    "Bearer",
		}))
	}
}

// Register – POST /register (public, no auth)
// @Summary      สมัครสมาชิกผู้ใช้ใหม่
// @Description  ลงทะเบียนผู้ใช้ทั่วไป (ไม่ต้องใช้ token)
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        user body presenter.UserCreate true "ข้อมูลสมัคร"
// @Success      200  {object}  map[string]interface{}
// @Failure      400  {object}  responses.ErrorResponse
// @Failure      422  {object}  responses.ErrorResponse
// @Router       /register [post]
func (h *userHandler) Register() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		req := new(presenter.UserCreate)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		newUser, err := h.usersUC.CreateUser(r.Context(), mapModel(req), req.ConfirmPassword)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		userResp := mapModelResponse(newUser)
		if userResp == nil {
			render.Render(w, r, responses.CreateErrorResponse(errors.New("internal server error")))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(*userResp))
	}
}

// Create – POST /user (admin only)
// @Security      BearerAuth
// @Summary       Create user (admin only)
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         user body presenter.UserCreate true "User creation data"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Failure       403  {object}  responses.ErrorResponse
// @Router        /user [post]
func (h *userHandler) Create() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		req := new(presenter.UserCreate)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			h.logger.Errorf("decode request body error: %v", err)
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		h.logger.Infof("request body decoded: email=%s, role_id=%d", req.Email, req.RoleID)

		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			h.logger.Warnf("validation failed: %v", err)
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		h.logger.Info("validation passed")

		userModel := mapModel(req)
		if userModel == nil {
			h.logger.Error("mapModel returned nil")
			render.Render(w, r, responses.CreateErrorResponse(errors.New("internal mapping error")))
			return
		}
		h.logger.Infof("mapped to model: email=%s, role_id=%d, status=%d", userModel.Email, userModel.RoleID, userModel.Status)

		newUser, err := h.usersUC.CreateUser(r.Context(), userModel, req.ConfirmPassword)
		if err != nil {
			h.logger.Errorf("CreateUser error: %v", err)
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		h.logger.Infof("user created successfully with id=%s", newUser.ID)

		responseData := mapModelResponse(newUser)
		if responseData == nil {
			h.logger.Error("mapModelResponse returned nil")
			render.Render(w, r, responses.CreateErrorResponse(errors.New("internal response error")))
			return
		}

		render.Respond(w, r, responses.CreateSuccessResponse(responseData))
	}
}

// Get – GET /user/{id}
// @Security      BearerAuth
// @Summary       Get user by ID
// @Tags          users
// @Produce       json
// @Param         id path string true "User ID"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Failure       404  {object}  responses.ErrorResponse
// @Router        /user/{id} [get]
func (h *userHandler) Get() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		user, err := h.usersUC.Get(r.Context(), id)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(user)))
	}
}

// GetMulti – GET /user?limit=10&offset=0
// @Security      BearerAuth
// @Summary       List users (admin only)
// @Tags          users
// @Produce       json
// @Param         limit query int false "Page size" default(20)
// @Param         offset query int false "Offset" default(0)
// @Param         email query string false "Filter by email"
// @Param         username query string false "Filter by username"
// @Param         status query int false "Filter by status"
// @Param         role_id query int false "Filter by role ID"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user [get]
func (h *userHandler) GetMulti() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		q := r.URL.Query()
		limit, _ := strconv.Atoi(q.Get("limit"))
		if limit <= 0 {
			limit = 20
		}
		offset, _ := strconv.Atoi(q.Get("offset"))
		if offset < 0 {
			offset = 0
		}

		filters := make(map[string]interface{})
		if email := q.Get("email"); email != "" {
			filters["email"] = email
		}
		if username := q.Get("username"); username != "" {
			filters["username"] = username
		}
		if fullname := q.Get("fullname"); fullname != "" {
			filters["fullname"] = fullname
		}
		if statusStr := q.Get("status"); statusStr != "" {
			if status, err := strconv.ParseInt(statusStr, 10, 16); err == nil {
				filters["status"] = int16(status)
			}
		}
		if roleStr := q.Get("role_id"); roleStr != "" {
			if roleID, err := strconv.Atoi(roleStr); err == nil {
				filters["role_id"] = roleID
			}
		}
		if verifiedStr := q.Get("verified"); verifiedStr != "" {
			filters["verified"] = verifiedStr == "true"
		}
		if sortBy := q.Get("sort_by"); sortBy != "" {
			filters["sort_by"] = sortBy
		}
		if sortOrder := q.Get("sort_order"); sortOrder != "" {
			filters["sort_order"] = sortOrder
		}

		users, total, err := h.usersUC.GetMultiWithTotal(r.Context(), limit, offset, filters)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}

		response := map[string]interface{}{
			"payload": mapModelsResponse(users),
			"pagination": map[string]interface{}{
				"limit":  limit,
				"offset": offset,
				"total":  total,
				"next_offset": func() int {
					if offset+limit < int(total) {
						return offset + limit
					}
					return -1
				}(),
			},
		}
		render.Respond(w, r, responses.CreateSuccessResponse(response))
	}
}

// Delete – DELETE /user/{id} (admin only)
// @Security      BearerAuth
// @Summary       Delete user (admin only)
// @Tags          users
// @Param         id path string true "User ID"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Failure       404  {object}  responses.ErrorResponse
// @Router        /user/{id} [delete]
func (h *userHandler) Delete() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		user, err := h.usersUC.Delete(r.Context(), id)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(user)))
	}
}

// Update – PUT /user/{id} (admin only)
// @Security      BearerAuth
// @Summary       Update user (admin only)
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         id path string true "User ID"
// @Param         user body presenter.UserUpdate true "Fields to update"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/{id} [put]
func (h *userHandler) Update() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		req := new(presenter.UserUpdate)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		values := make(map[string]interface{})
		if req.Firstname != nil {
			values["firstname"] = *req.Firstname
		}
		if req.Lastname != nil {
			values["lastname"] = *req.Lastname
		}
		if req.Fullname != nil {
			values["fullname"] = *req.Fullname
		}
		if req.MobileNumber != nil {
			values["mobile_number"] = *req.MobileNumber
		}
		if req.PhoneNumber != nil {
			values["phone_number"] = *req.PhoneNumber
		}
		if req.LineID != nil {
			values["lineid"] = *req.LineID
		}
		if req.LocationID != nil {
			values["location_id"] = *req.LocationID
		}
		updatedUser, err := h.usersUC.Update(r.Context(), id, values)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(updatedUser)))
	}
}

// UpdatePassword – PATCH /user/{id}/updatepass (admin only)
// @Security      BearerAuth
// @Summary       Update user password (admin only)
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         id path string true "User ID"
// @Param         password body presenter.UserUpdatePassword true "Password data"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/{id}/updatepass [patch]
func (h *userHandler) UpdatePassword() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		req := new(presenter.UserUpdatePassword)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		updatedUser, err := h.usersUC.UpdatePassword(r.Context(), id, req.OldPassword, req.NewPassword, req.ConfirmPassword)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(updatedUser)))
	}
}

// UpdateRole – PATCH /user/{id}/role (admin only)
// @Security      BearerAuth
// @Summary       Update user role (admin only)
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         id path string true "User ID"
// @Param         role body presenter.UserUpdateRole true "Role ID"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/{id}/role [patch]
func (h *userHandler) UpdateRole() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		req := new(presenter.UserUpdateRole)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		if req.RoleID <= 0 {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(errors.New("invalid role id"))))
			return
		}
		updatedUser, err := h.usersUC.Update(r.Context(), id, map[string]interface{}{"role_id": req.RoleID})
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(updatedUser)))
	}
}

// Me – GET /user/me
// @Security      BearerAuth
// @Summary       Get current user profile
// @Tags          users
// @Produce       json
// @Success       200  {object}  map[string]interface{}
// @Failure       401  {object}  responses.ErrorResponse
// @Router        /user/me [get]
func (h *userHandler) Me() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, err := middleware.GetUserFromCtx(r.Context())
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(user)))
	}
}

// UpdateMe – PUT /user/me
// @Security      BearerAuth
// @Summary       Update current user profile
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         user body presenter.UserUpdate true "Fields to update"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/me [put]
func (h *userHandler) UpdateMe() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, err := middleware.GetUserFromCtx(r.Context())
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		req := new(presenter.UserUpdate)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		values := make(map[string]interface{})
		if req.Firstname != nil {
			values["firstname"] = *req.Firstname
		}
		if req.Lastname != nil {
			values["lastname"] = *req.Lastname
		}
		if req.Fullname != nil {
			values["fullname"] = *req.Fullname
		}
		if req.MobileNumber != nil {
			values["mobile_number"] = *req.MobileNumber
		}
		if req.PhoneNumber != nil {
			values["phone_number"] = *req.PhoneNumber
		}
		if req.LineID != nil {
			values["lineid"] = *req.LineID
		}
		if req.LocationID != nil {
			values["location_id"] = *req.LocationID
		}
		updatedUser, err := h.usersUC.Update(r.Context(), user.ID, values)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(updatedUser)))
	}
}

// UpdatePasswordMe – PATCH /user/me/updatepass
// @Security      BearerAuth
// @Summary       Change current user password
// @Tags          users
// @Accept        json
// @Produce       json
// @Param         password body presenter.UserUpdatePassword true "Password data"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/me/updatepass [patch]
func (h *userHandler) UpdatePasswordMe() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		user, err := middleware.GetUserFromCtx(r.Context())
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		req := new(presenter.UserUpdatePassword)
		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		if err := utils.ValidateStruct(r.Context(), req); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		updatedUser, err := h.usersUC.UpdatePassword(r.Context(), user.ID, req.OldPassword, req.NewPassword, req.ConfirmPassword)
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(mapModelResponse(updatedUser)))
	}
}

// LogoutAllAdmin – GET /user/{id}/logoutall (admin only)
// @Security      BearerAuth
// @Summary       Force logout all sessions of a user (admin only)
// @Tags          users
// @Param         id path string true "User ID"
// @Success       200  {object}  map[string]interface{}
// @Failure       400  {object}  responses.ErrorResponse
// @Router        /user/{id}/logoutall [get]
func (h *userHandler) LogoutAllAdmin() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id, err := uuid.Parse(chi.URLParam(r, "id"))
		if err != nil {
			render.Render(w, r, responses.CreateErrorResponse(httpErrors.ErrValidation(err)))
			return
		}
		if err := h.usersUC.LogoutAll(r.Context(), id); err != nil {
			render.Render(w, r, responses.CreateErrorResponse(err))
			return
		}
		render.Respond(w, r, responses.CreateSuccessResponse(struct{}{}))
	}
}

// ------------------------------
// Mapping functions
// ------------------------------

func mapModel(req *presenter.UserCreate) *models.SdUser {
	var fullname *string
	if req.Fullname != "" {
		fullname = &req.Fullname
	} else if req.Firstname != "" || req.Lastname != "" {
		combined := strings.TrimSpace(req.Firstname + " " + req.Lastname)
		if combined != "" {
			fullname = &combined
		}
	}
	return &models.SdUser{
		Email:        strings.ToLower(strings.TrimSpace(req.Email)),
		Username:     strings.ToLower(strings.TrimSpace(req.Username)),
		Password:     req.Password,
		PasswordTemp: stringPtr(req.Password),
		RoleID:       req.RoleID,
		Firstname:    stringPtr(req.Firstname),
		Lastname:     stringPtr(req.Lastname),
		Fullname:     fullname,
		MobileNumber: stringPtr(req.MobileNumber),
		PhoneNumber:  stringPtr(req.PhoneNumber),
		LineID:       stringPtr(req.LineID),
		LocationID:   stringPtr(req.LocationID),
		Status:       1,
		IsSuperUser:  false,
		Verified:     false,
	}
}

func mapModelResponse(user *models.SdUser) *presenter.UserResponse {
	if user == nil {
		return nil
	}
	return &presenter.UserResponse{
		ID:           user.ID,
		Email:        user.Email,
		Username:     user.Username,
		RoleID:       user.RoleID,
		Firstname:    user.Firstname,
		Lastname:     user.Lastname,
		Fullname:     user.Fullname,
		Nickname:     user.Nickname,
		MobileNumber: user.MobileNumber,
		PhoneNumber:  user.PhoneNumber,
		LineID:       user.LineID,
		LocationID:   user.LocationID,
		Avatar:       user.Avatar,
		Gender:       user.Gender,
		Birthday:     user.Birthday,
		Status:       user.Status,
		IsSuperUser:  user.IsSuperUser,
		Verified:     user.Verified,
		LastSignIn:   &user.Lastsignindate,
		Remark:       user.Remark,
		CreatedAt:    user.CreatedDate,
		UpdatedAt:    user.UpdatedDate,
	}
}

func mapModelsResponse(users []*models.SdUser) []*presenter.UserResponse {
	out := make([]*presenter.UserResponse, len(users))
	for i, u := range users {
		out[i] = mapModelResponse(u)
	}
	return out
}

func stringPtr(s string) *string {
	if s == "" {
		return nil
	}
	return &s
}